requirements.txt
pandas
scikit-learn
joblib
gradio
PyGithub